+++ changelog.md
@@ -141,3 +141,7 @@ All notable changes to `PagesModule` will be documented in this file.
 
 ## Version 6.0.2
 - Bugfix: admin page preview
+
+## Version 6.0.3
+- Bugfix: page title
+- Bugfix: change image in gallery